backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
PyJWT>=2.9,<3.0
passlib>=1.7,<2.0
python-multipart>=0.0.9,<1.0
